also get ratings for search results
