+++ src/services/resource/ratedresources.hooks.js
@@ -1,21 +1,6 @@
 const validateResourceSchema = require('../../hooks/validate-resource-schema/');
 const authenticate = require('../../hooks/authenticate');
-
-const getRating = context => function(resource) {
-  return context.app.service('ratings').Model.aggregate([
-    {$match: {materialId: resource._id}},
-    {$group: {_id: "$isTeacherRating", avgRating: {$avg: "$rating"}}}
-  ]).then(aggregatedRatings => {
-    //TODO X make aggregate return the results in a easier to use format (using $project?)
-    aggregatedRatings.forEach(it => {
-      if (it._id === true) {
-        resource.teacherRating = it.avgRating;
-      } else if (it._id === false) {
-        resource.studentRating = it.avgRating;
-      }
-    });
-  })
-};
+const getRatingHook = require('../../hooks/get-rating-hook');
 
 module.exports = {
   before: {
@@ -30,14 +15,8 @@ module.exports = {
 
   after: {
     all: [],
-    find: [
-      context =>
-        Promise.all(context.result.data.map(getRating(context))).then(() => context)
-    ],
-    get: [
-      context =>
-        getRating(context)(context.result).then(() => context)
-    ],
+    find: [getRatingHook.find],
+    get: [getRatingHook.get],
     create: [],
     update: [],
     patch: [],
